requirements.txt
Django==5.0.1
fontawesomefree
Pillow
django-import-export==3.2.0
djangoql
python-dotenv